requirements.txt
streamlit==1.29.0
pymupdf==1.23.8
pinecone-client==2.2.4
langchain==0.0.347
openai==0.28.1
moviepy==1.0.3
librosa==0.10.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
transformers==4.35.0
requests==2.31.0
sqlalchemy==2.0.23
fastapi
uvicorn
python-multipart